# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HealthTrack API
  version: 1.0.0
  description: |
    Unified healthcare record system API enabling users to upload, parse, verify, 
    or manually enter prescriptions & lab reports.
  contact:
    name: HealthTrack Team
    email: api@healthtrack.io
servers:
  - url: https://api.healthtrack.io/v1
    description: Production API
  - url: https://api-dev.healthtrack.io/v1
    description: Development API
security:
  - BearerAuth: []
paths:
...
  /records/{recordId}:
    get:
      summary: Get record details
      description: Retrieve detailed information for a specific record
      operationId: getRecord
      tags:
        - Records
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Record retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
  schemas:
    Record:
      type: object
      required:
        - recordId
        - userId
        - recordType
        - source
        - verified
        - data
        - createdAt
        - updatedAt
      properties:
        recordId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        userId:
          type: string
          example: "U123"
        recordType:
          type: string
          enum: [labReport, course]
        source:
          type: string
          enum: [manual, ocr, api]
        verified:
          type: boolean
        version:
          type: integer
          default: 1
        data:
          type: object
          description: Dynamic data based on recordType
        mcpContextURI:
          type: string
          format: uri
          example: "https://api.healthtrack.io/v1/context/lab-report:U123:LR456"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...